requirements.txt
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.1
Flask==3.1.2
fonttools==4.61.1
fsspec==2025.12.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.0
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.2.6
opencv-python==4.12.0.88
openvino==2025.4.1
openvino-telemetry==2025.2.0
packaging==25.0
pillow==12.0.0
polars==1.36.1
polars-runtime-32==1.36.1
psutil==7.2.0
pyparsing==3.3.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
RPi.GPIO==0.7.1
scipy==1.16.3
six==1.17.0
sympy==1.14.0
torch==2.9.1+cpu
torchvision==0.24.1
typing_extensions==4.15.0
ultralytics==8.3.241
ultralytics-thop==2.0.18
urllib3==2.6.2
Werkzeug==3.1.4
